[WFLY-18499] Update the microprofile-rest-client for the Common Enhancements. Refactor the quickstart to a single module with a test for the MicroProfile REST Client.
