+++ microprofile-rest-client/charts/helm.yaml
@@ -1,0 +1,6 @@
+build:
+  uri: https://github.com/wildfly/quickstart.git
+  ref: main
+  contextDir: microprofile-rest-client
+deploy:
+  replicas: 1
