--- microprofile-rest-client/country-client/pom.xml
@@ -1,163 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <groupId>org.wildfly.quickstarts</groupId>
-    <artifactId>microprofile-rest-client</artifactId>
-    <!--
-    Maintain separation between the artifact id and the version to help prevent
-    merge conflicts between commits changing the GA and those changing the V.
-    -->
-    <version>31.0.0.Beta1-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>country-client</artifactId>
-  <packaging>war</packaging>
-
-  <name>Quickstart: microprofile-rest-client/country-client</name>
-
-  <dependencies>
-    <!-- Import the MicroProfile REST Client API, we use provided scope as the API is included in the server -->
-    <dependency>
-      <groupId>org.eclipse.microprofile.rest.client</groupId>
-      <artifactId>microprofile-rest-client-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
-    <!-- Import the MicroProfile Config API, we use provided scope as the API is included in the server -->
-    <dependency>
-      <groupId>org.eclipse.microprofile.config</groupId>
-      <artifactId>microprofile-config-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
-    <!-- Import the CDI API, we use provided scope as the API is included in the server -->
-    <dependency>
-      <groupId>jakarta.enterprise</groupId>
-      <artifactId>jakarta.enterprise.cdi-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
-    <!-- Import the Jakarta REST API, we use provided scope as the API is included in the server -->
-    <dependency>
-      <groupId>jakarta.ws.rs</groupId>
-      <artifactId>jakarta.ws.rs-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
-    <!-- Needed for running tests (you may also use TestNG) -->
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <!-- Arquillian allows you to test enterprise code such as EJBs and
-           Transactional(JTA) JPA from JUnit/TestNG -->
-    <dependency>
-      <groupId>org.jboss.arquillian.junit</groupId>
-      <artifactId>arquillian-junit-container</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.wildfly.arquillian</groupId>
-      <artifactId>wildfly-arquillian-common</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.resteasy</groupId>
-      <artifactId>resteasy-client</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.logging</groupId>
-      <artifactId>commons-logging-jboss-logging</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <!-- Set the name of the archive -->
-    <finalName>${project.artifactId}</finalName>
-  </build>
-
-  <profiles>
-    <profile>
-      <id>bootable-jar</id>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.wildfly.plugins</groupId>
-            <artifactId>wildfly-jar-maven-plugin</artifactId>
-            <configuration>
-              <feature-pack-location>wildfly@maven(org.jboss.universe:community-universe)#${version.server.bootable-jar}</feature-pack-location>
-              <layers>
-                <layer>jaxrs-server</layer>
-                <layer>microprofile-platform</layer>
-              </layers>
-              <plugin-options>
-                <jboss-fork-embedded>true</jboss-fork-embedded>
-              </plugin-options>
-            </configuration>
-            <executions>
-              <execution>
-                <goals>
-                  <goal>package</goal>
-                </goals>
-              </execution>
-            </executions>
-          </plugin>
-        </plugins>
-      </build>
-    </profile>
-    <profile>
-      <id>bootable-jar-openshift</id>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.wildfly.plugins</groupId>
-            <artifactId>wildfly-jar-maven-plugin</artifactId>
-            <configuration>
-              <feature-pack-location>wildfly@maven(org.jboss.universe:community-universe)#${version.server.bootable-jar}</feature-pack-location>
-              <layers>
-                <layer>jaxrs-server</layer>
-                <layer>microprofile-platform</layer>
-              </layers>
-              <plugin-options>
-                <jboss-fork-embedded>true</jboss-fork-embedded>
-              </plugin-options>
-              <cloud/>
-            </configuration>
-            <executions>
-              <execution>
-                <goals>
-                  <goal>package</goal>
-                </goals>
-              </execution>
-            </executions>
-          </plugin>
-          <plugin>
-            <groupId>org.eclipse.jkube</groupId>
-            <artifactId>openshift-maven-plugin</artifactId>
-            <executions>
-              <execution>
-                <goals>
-                  <goal>resource</goal>
-                  <goal>build</goal>
-                </goals>
-              </execution>
-            </executions>
-            <configuration>
-              <enricher>
-                <config>
-                  <jkube-service>
-                    <type>NodePort</type>
-                  </jkube-service>
-                </config>
-              </enricher>
-            </configuration>
-          </plugin>
-        </plugins>
-      </build>
-    </profile>
-  </profiles>
-
-</project>
