--- microprofile-rest-client/country-client/src/main/java/org/wildfly/quickstarts/microprofile/rest/client/JaxRsApplication.java
@@ -1,8 +1,0 @@
-package org.wildfly.quickstarts.microprofile.rest.client;
-
-import jakarta.ws.rs.ApplicationPath;
-import jakarta.ws.rs.core.Application;
-
-@ApplicationPath("/")
-public class JaxRsApplication extends Application {
-}
