--- microprofile-rest-client/country-client/src/main/java/org/wildfly/quickstarts/microprofile/rest/client/RootResource.java
@@ -1,16 +1,0 @@
-package org.wildfly.quickstarts.microprofile.rest.client;
-
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-
-@Path("/")
-public class RootResource {
-
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String getRootResponse() {
-        return "MicroProfile REST Client quickstart deployed successfully. You can find the available operations in the included README file.";
-    }
-}
