--- microprofile-rest-client/country-client/src/main/jkube/deployment.yml
@@ -1,7 +1,0 @@
-spec:
-  template:
-    spec:
-      containers:
-        - env:
-            - name: JAVA_OPTS_APPEND
-              value: "-Dorg.wildfly.quickstarts.microprofile.rest.client.CountriesServiceClient/mp-rest/url=http://country-server:8080"
