--- microprofile-rest-client/country-client/src/main/jkube/route.yml
@@ -1,7 +1,0 @@
-spec:
-  tls:
-    insecureEdgeTerminationPolicy: Redirect
-    termination: edge
-  to:
-    kind: Service
-    name: country-client
