--- microprofile-rest-client/country-client/src/main/resources/META-INF/microprofile-config.properties
@@ -1,2 +1,0 @@
-org.wildfly.quickstarts.microprofile.rest.client.CountriesServiceClient/mp-rest/url=http://localhost:8080/country-server
-org.wildfly.quickstarts.microprofile.rest.client.CountriesServiceClient/mp-rest/scope=jakarta.inject.Singleton
