--- microprofile-rest-client/country-server/src/main/java/org/wildfly/quickstarts/microprofile/JaxRsApplication.java
@@ -1,8 +1,0 @@
-package org.wildfly.quickstarts.microprofile;
-
-import jakarta.ws.rs.ApplicationPath;
-import jakarta.ws.rs.core.Application;
-
-@ApplicationPath("/")
-public class JaxRsApplication extends Application {
-}
