+++ microprofile-rest-client/src/main/java/org/wildfly/quickstarts/microprofile/Country.java
@@ -1,0 +1,35 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ *
+ * Copyright 2023 Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wildfly.quickstarts.microprofile;
+
+public class Country {
+
+    public String name;
+    public String capital;
+    public String currency;
+
+    public Country(){}
+
+    public Country(String name, String capital, String currency) {
+        this.name = name;
+        this.capital = capital;
+        this.currency = currency;
+    }
+}
