+++ microprofile-rest-client/src/test/java/org/wildfly/quickstarts/microprofile/rest/client/NotFoundResponseExceptionMapper.java
@@ -1,0 +1,39 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ *
+ * Copyright 2023 Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wildfly.quickstarts.microprofile.rest.client;
+
+import org.eclipse.microprofile.rest.client.ext.ResponseExceptionMapper;
+
+import jakarta.ws.rs.NotFoundException;
+import jakarta.ws.rs.core.MultivaluedMap;
+import jakarta.ws.rs.core.Response;
+
+public class NotFoundResponseExceptionMapper implements ResponseExceptionMapper<NotFoundException> {
+
+    @Override
+    public boolean handles(int status, MultivaluedMap<String, Object> headers) {
+        return status == 404;
+    }
+
+    @Override
+    public NotFoundException toThrowable(Response response) {
+        return new NotFoundException(response.hasEntity() ? response.readEntity(String.class) : "Resource not found");
+    }
+}
