+++ microprofile-rest-client/src/test/resources/META-INF/microprofile-config.properties
@@ -1,0 +1,2 @@
+org.wildfly.quickstarts.microprofile.rest.client.CountriesServiceClient/mp-rest/url=${server.host}
+org.wildfly.quickstarts.microprofile.rest.client.CountriesServiceClient/mp-rest/scope=jakarta.inject.Singleton
