+++ shared-doc/xp-deploy-project.adoc
@@ -105,8 +105,8 @@ ifdef::rest-client-qs[]
 [options="nowrap",subs="+attributes"]
 ----
 $ oc new-app --template={ImagePrefixVersion}-basic-s2i {backslash}
- -p EAP_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk8-openshift:latest {backslash}
- -p EAP_RUNTIME_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk8-runtime-openshift:latest {backslash}
+ -p EAP_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk11-openshift:latest {backslash}
+ -p EAP_RUNTIME_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk11-runtime-openshift:latest {backslash}
  -p IMAGE_STREAM_NAMESPACE="{artifactId}-project" {backslash}
  -p SOURCE_REPOSITORY_URL="{EAPQuickStartRepo}" {backslash}
  -p SOURCE_REPOSITORY_REF="{EAPQuickStartRepoRef}" {backslash}
@@ -114,103 +114,8 @@ ifdef::uses-h2[]
  -p ENABLE_GENERATE_DEFAULT_DATASOURCE="true" {backslash}
 endif::[]
  -p CONTEXT_DIR="{artifactId}" \
- -p ARTIFACT_DIR="country-server/target" \
+ -p ARTIFACT_DIR="target" \
  -p APPLICATION_NAME="country-server"
-----
-+
---
-* `--template` The template to use.
-* `-p IMAGE_STREAM_NAMESPACE` The latest images streams and templates xref:import_imagestreams_templates[were imported into the project's namespace], so you must specify the namespace of where to find the image stream. This is usually the OpenShift project's name.
-* `-p SOURCE_REPOSITORY_URL` The URL to the repository containing the application source code.
-* `-p SOURCE_REPOSITORY_REF` The Git repository reference to use for the source code. This can be a Git branch or tag reference.
-ifdef::uses-h2[]
-* `-p ENABLE_GENERATE_DEFAULT_DATASOURCE` Enables the default datasource, which is required by this quickstart.
-endif::[]
-* `-p CONTEXT_DIR` The directory within the source repository to build.
-* `-P ARTIFACT_DIR` The directory from which to copy the artifact.
-* `-p APPLICATION_NAME` The name of the application.
---
-+
-Alternatively, to create the *country-server* application using the JDK 11 images enter the following command.
-+
-[options="nowrap",subs="+attributes"]
-----
-$ oc new-app --template={ImagePrefixVersion}-basic-s2i {backslash}
- -p EAP_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk11-openshift:latest \
- -p EAP_RUNTIME_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk11-runtime-openshift:latest \
- -p IMAGE_STREAM_NAMESPACE="{artifactId}-project" {backslash}
- -p SOURCE_REPOSITORY_URL="{EAPQuickStartRepo}" {backslash}
- -p SOURCE_REPOSITORY_REF="{EAPQuickStartRepoRef}" {backslash}
-ifdef::uses-h2[]
- -p ENABLE_GENERATE_DEFAULT_DATASOURCE="true" {backslash}
-endif::[]
- -p CONTEXT_DIR="{artifactId}" \
- -p ARTIFACT_DIR="country-server/target" \
- -p APPLICATION_NAME="country-server"
-----
-+
---
-* `--template` The template to use.
-* `-p IMAGE_STREAM_NAMESPACE` The latest images streams and templates xref:import_imagestreams_templates[were imported into the project's namespace], so you must specify the namespace of where to find the image stream. This is usually the OpenShift project's name.
-* `-p SOURCE_REPOSITORY_URL` The URL to the repository containing the application source code.
-* `-p SOURCE_REPOSITORY_REF` The Git repository reference to use for the source code. This can be a Git branch or tag reference.
-ifdef::uses-h2[]
-* `-p ENABLE_GENERATE_DEFAULT_DATASOURCE` Enables the default datasource, which is required by this quickstart.
-endif::[]
-* `-p CONTEXT_DIR` The directory within the source repository to build.
-* `-P ARTIFACT_DIR` The directory from which to copy the artifact.
-* `-p APPLICATION_NAME` The name of the application.
---
-+
-*Country-client application:*
-+
-[options="nowrap",subs="+attributes"]
-----
-$ oc new-app --template={ImagePrefixVersion}-basic-s2i {backslash}
- -p EAP_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk8-openshift:latest {backslash}
- -p EAP_RUNTIME_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk8-runtime-openshift:latest {backslash}
- -p IMAGE_STREAM_NAMESPACE="{artifactId}-project" {backslash}
- -p SOURCE_REPOSITORY_URL="{EAPQuickStartRepo}" {backslash}
- -p SOURCE_REPOSITORY_REF="{EAPQuickStartRepoRef}" {backslash}
-ifdef::uses-h2[]
- -p ENABLE_GENERATE_DEFAULT_DATASOURCE="true" {backslash}
-endif::[]
- -p CONTEXT_DIR="microprofile-rest-client" \
- -p ARTIFACT_DIR="country-client/target" \
- -p APPLICATION_NAME="country-client" \
- -e JAVA_OPTS_APPEND="-Dorg.wildfly.quickstarts.microprofile.rest.client.CountriesServiceClient/mp-rest/url=http://country-server:8080"
-----
-+
---
-* `--template` The template to use.
-* `-p IMAGE_STREAM_NAMESPACE` The latest images streams and templates xref:import_imagestreams_templates[were imported into the project's namespace], so you must specify the namespace of where to find the image stream. This is usually the OpenShift project's name.
-* `-p SOURCE_REPOSITORY_URL` The URL to the repository containing the application source code.
-* `-p SOURCE_REPOSITORY_REF` The Git repository reference to use for the source code. This can be a Git branch or tag reference.
-ifdef::uses-h2[]
-* `-p ENABLE_GENERATE_DEFAULT_DATASOURCE` Enables the default datasource, which is required by this quickstart.
-endif::[]
-* `-p CONTEXT_DIR` The directory within the source repository to build.
-* `-P ARTIFACT_DIR` The directory from which to copy the artifact.
-* `-p APPLICATION_NAME` The name of the application.
---
-+
-Alternatively, to create the *country-client* application using the JDK 11 images enter the following command.
-+
-[options="nowrap",subs="+attributes"]
-----
-$ oc new-app --template={ImagePrefixVersion}-basic-s2i {backslash}
- -p EAP_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk11-openshift:latest \
- -p EAP_RUNTIME_IMAGE_NAME=jboss-{ImagePrefixVersion}-openjdk11-runtime-openshift:latest \
- -p IMAGE_STREAM_NAMESPACE="{artifactId}-project" {backslash}
- -p SOURCE_REPOSITORY_URL="{EAPQuickStartRepo}" {backslash}
- -p SOURCE_REPOSITORY_REF="{EAPQuickStartRepoRef}" {backslash}
-ifdef::uses-h2[]
- -p ENABLE_GENERATE_DEFAULT_DATASOURCE="true" {backslash}
-endif::[]
- -p CONTEXT_DIR="microprofile-rest-client" \
- -p ARTIFACT_DIR="country-client/target" \
- -p APPLICATION_NAME="country-client" \
- -e JAVA_OPTS_APPEND="-Dorg.wildfly.quickstarts.microprofile.rest.client.CountriesServiceClient/mp-rest/url=http://country-server:8080"
 ----
 +
 --
